requirements.txt
google-cloud-secret-manager==2.16.3
redis==5.*
requests==2.31.0
requests-oauthlib==1.3.1
Flask==2.3.3
